APPS-8616: Deprecation warning for "$defaultFloat"
Added !global to the declarations that were generating the warning, as indicated by the instructions in the warning:

DEPRECATION WARNING on line 43 of ~/.rvm/gems/ruby-2.2.4@paynearme4/bundler/gems/foundation-a7ead9b6be09/scss/foundation/_settings.scss:
Assigning to global variable "$defaultFloat" by default is deprecated.
In future versions of Sass, this will create a new local variable.
If you want to assign to the global variable, use "$defaultFloat: left !global" instead.
Note that this will be incompatible with Sass 3.2.